Expose interpreter state for interpreters with transformers (#2005)
If an interpreter is configured with a transformer, the `interpreter_state` admin endpoint is not available for that interpreter.  Additionally, if interpreters are configured on multiple routers, each interpreter will attempt to register a handler at the same URL and override the others.

We update transformers to pass through the admin handlers of the underlying interpreter so that the `interpreter_state` remains available.  We also update the URL structure of the `interpreter_state` endpoint to be prefixed with the router name.  This avoids any cross-router conflicts.

Fixes #1999 
Fixes #2000

Signed-off-by: Alex Leong <alex@buoyant.io>