use custom RetryFilter to retry consul connection refused exceptions (#2024)
When Linkerd is configured to use `io.l5d.consul` and it is unable to connect to consul because of a `ConnectionRefused` exception, it is supposed to retry the connection infinitely. Currently, after Linkerd makes an attempt to the local consul agent, it fails to retry the connection. This is because Finagle's `RetryFilter` considers `ConnectionRefused` exceptions as `NonRetryable`.

This PR uses a custom `RetryFilter` to retry all HTTP interactions to a consul. The custom filter is not restrictive in its determination of what exceptions are retryable and only relies on the `RetryPolicy` configured for that filter.

Tests were conducted locally with the help of @aywrite's [repo](https://github.com/aywrite/linkerd-consul-test) and a locally running consul agent.

fixes #2003

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>