Upgrade to Finagle 18.5.0 and switch from Buf to ByteBuf (#2020)
Upgrade to Finagle 18.5.0.  In this version of Finagle, certain utility methods for converting between Buf and ByteBuf were removed.  This gives us the perfect opportunity to change the type of the `buf` field of `Frame.Data` from Buf to ByteBuf.

Netty uses ByteBuf as its byte buffer type throughout.  In our H2 implementation we wrap Netty's ByteBufs in Buf, Finagle's byte buffer type.  Buf is type that provides a nice API and abstracts over a variety of byte buffer types such as Array[Byte], ByteBuffer, and ByteBuf.  However, since we only use byte buffers from Netty, our Bufs always contain a Netty ByteBuf.  Not only does this make the abstraction unnecessary but the abstraction is also leaky because we require that the Bufs contain a ByteBuf in order for Netty's ByteBuf reference counting to work correctly.

We change `Frame.Data` to hold a Netty ByteBuf directly instead of a Buf which wraps a ByteBuf.  This avoids the need to convert between the two and eliminates a pointless abstraction.

Signed-off-by: Alex Leong <alex@buoyant.io>