Don't track remote reset streams (#2047)
When a stream reset is received from the remote, the Netty4Dispatcher tracks the state of that stream as StreamRemoteReset.  Streams in this state are treated exactly the same as closed streams but are never removed from the streams map.  This leads to a memory leak on long lived connections if many streams are remote reset.

Since remote reset streams are treated exactly the same as closed streams, we can simply remove these streams from the steams map instead of tracking their state.

Signed-off-by: Alex Leong <alex@buoyant.io>