Fix flakey retry test (#2059)
RetriesEndToEndTest sometimes fails.  I believe this is because of a race condition where the time control can be advanced before Future.sleep is called.  Because the time control is never advanced again, this Future stays pending forever.

Enlarge the scope of the time control block to make the time control available to the downstream service.  Only advance the time control after Future.sleep is called.

Because of the nature of flaky tests, it is hard to verify that this is fixed.  I manually ran this test 20 times after applying this fix and did not see it fail.

Fixes #1962 (probably)

Signed-off-by: Alex Leong <alex@buoyant.io>