Pass router param inputs to identifiers (#2056)
Stack params are passed to identifier plugins to configure them.  However, only params from the router config are passed to the identifier.  In particular, this means that the stats receiver is not available to the identifier because it is not a router config param.

Pass stack params as input to the router params.  This allows the stats receiver to be passed into the router params.

Fixes #1963 

Signed-off-by: Alex Leong <alex@buoyant.io>