Always clear the x-forwarded-client-cert header (#2085)
When Linkerd receives a x-forwarded-client-cert header field in a request, it must remove it or it must reject the request. It must never forward the x-forwarded-client-cert header field because that would allow any client to forge the identity of other clients if the inner application is relying on the x-forwarded-client-cert header field. (Also some applications may look at the x-forwarded-client-cert to determine if the request was sent over HTTPS or HTTP.)

Always clear any incoming x-forwarded-client-cert header.

Fixes #2068 

Signed-off-by: Alex Leong <alex@buoyant.io>