[WIP] Add an io.l5d.destination interface to support Linkerd2 (#2091)
There is no set way for Linkerd2 proxies to use Namerd service discovery in a Kubernetes daemonset environment. Linkerd 2 uses its control plane's gRPC destination API, while Linkerd 1.x uses Namerd for its service discovery and control plane. To enable a Linkerd2-proxy to work alongside Namerd and Linkerd 1.x, Namerd needs to implement the gRPC destination API and use its namers to provide service discovery information to an l5d2 proxy.

This PR adds a new Namerd io.l5d.destination interface that implements the API and resolves services names using configured namers. When a destination request reaches the interface, a gRPC stream is created and Namerd creates an Activity that interprets a given destination path into a set of addresses. When a stream is created, the interface immediately sends a NoEndpoints message followed by an Add message if there are addresses available for a service name. As addresses are discovered and/or removed, the interface calculates the delta between previously discovered and the current address and sends Add and Remove gRPC messages back to the client. If there is no address change, i.e., if there is nothing to add or remove, we assume that the service associated with that service exists but does not have endpoints available, and the interface sends a NoEndpoints message with exists: true.

The interface has been tested with controller/script/destination-client in the linkerd2 repo.

One major caveat with the io.l5d.destination is that does not provide reach resource labels that Linkerd2 uses for metrics. This means that metrics generated by a proxy backed Namerd could potentially be sparse and more not give in-depth request telemetry when compared to Linkerd2's destination API.