Add GC logging by default (#2119)
GC logging is necessary for debugging JVM issues in Linkerd. This PR adds flags to the Linkerd exec script to enable GC logging. It also adds a line in the exec script that creates a directory var/log/linkerd for gc.log files if it doesn't exist. it also allows for the GC log directory to be configurable using the environment variable GC_LOG

fixes #2110

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>