Add request timeout to consul client to avoid stale service discovery updates (#2140)
There have been issues where Namerd with a consul namer fails to update its watch state in situations were a long-polling request to consul never returns with service discovery info.

This PR adds a request timeout configuration for the client that issues these requests. The timeout is set to 10 minutes, which is double the timeout that consul already provides for blocking queries. When Linkerd's consul client reaches the 10 minute timeout, the client throws an IndividualRequestTimeoutException which triggers a new request to be issued in SvcAddr to the consul agent.

Tests were performed manually with a Test server that responded to any consul client request after 10 minutes.

fixes #2131

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>