removes all load balancers from the balancer registry (#2153)
Linkerd runs into an issue where it runs out of memory when it receives empty address sets after a while from any configured namer. This is due to Finagle's Balancer registry not managing balancer state correctly and inadvertently store balancer instances in its BalancerRegistry that are never removed.

This PR is set up to effectively "disable" the BalancerRegistry module by immediately removing Balancer instances when they are created. Once the issue is fixed we can remove this workaround.

Tests were performed with a locally running Linkerd instance backed by a fs namer. Removing and readding an ip address in the cat file caused the bug to appear. After this fix, the size of the BalancerRegistery does not grow. Tests were also done with all the various load balancer configurations with Linkerd.

fixes #2152

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>