stabilizeExistence's inner var should watch the underlying var (#2164)
Fixes #2079 

The `stabilizeExistence` method returns a type that consists of an outer `Var[Activity.State[Option[Var[T]]]]` which contains an inner `Var[T]`.  The outer `Var.async` which is driven by updates from the underlying unstable Activity.  This mostly behaves as intended with changes to the underlying unstable Activity causing updates to the outer and inner Vars as appropriate.  However, the inner Var is not an `Var.async` and is only updated so long as the outer Var is observed.  This means that if the inner Var is observed and the outer Var is not, changes will not propagate from the underlying unstable Activity to the inner Var.

This can cause a load balancer to get get stuck with an address set which never updates, even when new data comes in from service discovery.  Consider this scenario:

A first call to the Kubernetes namer is made and the namer calls `stabilizeExistence` on the underlying endpoints Activity.  The Activity (which is essentially the outer Var) is observed and the inner `Name.Bound` (which contains the inner Var) is cached as part of a client stack and also observed.  Then the observation on the outer Var is closed.  The inner Var is still observed but will not receive any updates from the endpoints Activity that backs it.  Since the client stack keeps this Var[Addr] cached, it will hold on to this stale Var for the lifetime of the client stack.

We solve this by making both the inner and outer Vars to be `Var.async` with each of them observing the underlying Activity.  This way, each Var will get updates if and only if that Var is observed.

Signed-off-by: Alex Leong <alex@buoyant.io>