Retain last state on "No path to datacenter" error
Consul responds with "No path to datacenter" when it is queried about DC
that does not exist (case 1) or DC that can't be contacted due to
network partition (case 2). Currently consul namer responds with
Addr.Neg on such response.
This makes sense in (case 1), because we want to fallback to
explore alternatives in dtab asap when we try to resolve new name.
However in (case 2) we want to retain previously known state (if it
exists), so namerd could survive intermittent consul failures caused by
network partition.

This commit adjusts consul namer so it only returns Addr.Neg if current
state is Addr.Pending. Otherwise, previous address is retained.

Fixes #2200

Signed-off-by: Dmytro Kostiuchenko <edio@archlinux.us>