add missing forward slash in linkerd/namerd exec script (#2213)
This commit fixes an issue where part of the exec script for Linkerd and Namerd are unable to check the java version due to a missing forward slash.

Tests were conducted in docker where the JAVA_HOME variable was set to an empty string in the docker image. Then both Linkerd and Namerd were run separately to exercise the offending branch of the bash script. Before the fix, the message line 24: usr/bin/java: No such file or directory: integer expression expected appeared in the logs. After this commit, the message is no longer visible.

Fixes #2207

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>