Issue 2212: Fix Namerd memory leak (#2222)
# Problem

Namerd's memory usage keeps increasing and will not perform garbage collection
until its usage hit 80%~100%. A possible consequence is that, if the GC cannot
be finished properly, the n4d pods will unable to serve any requests unless it's
restarted. A side fact is that, linkerd's memory usuage has ever gone up higher
than 20%, which might mean Namerd is not releasing the objects/resources
properly which caused a memory leak.

# Solution

namerd keeps open watches for every `getDtab` request that is made. When the value of that request is fulfilled, the watch for that request is kept in the `ZKWatcherManager` as seen in [1] after 5k requests sent.

namerd should remove the watch for a dtab once a value has been fulfilled. In order to do that, `def ApacheZooKeeper.getDataWatch` should handle interrupts for `FutureCancelledExceptions`. When it handles this interrupt, it now removes the watch from `ZKWatcherManager`. Adding this functionality required forking our own version of `ApacheZooKeper.scala` and is the reason for the addition of this file.

Once we are able to remove watches, `def ZkSession.watchedOperation` no longer should fire and forget its retries. This required refactoring how to handle `WatchedState` changes to `loop()` only in certain cases.

The result of these changes is seen in [2] after 5k requests sent.

# Validation

[1] and [2] show the changes reflected by this change. `ZKWatchManager` now properly removes watches when values for dtabs have been fulfilled. It continues to work as expected for streams.

Fixes #2212

Signed-off-by: Kevin Leimkuhler <kevinl@buoyant.io>
Co-authored-by: Alex Leong <alex@buoyant.io>

[1]
![screen shot 2019-02-15 at 12 00 06 pm](https://user-images.githubusercontent.com/4572153/52881383-40a4db00-3119-11e9-9eca-1bffd044d7c0.png)

[2]
![screen shot 2019-02-15 at 11 58 50 am](https://user-images.githubusercontent.com/4572153/52881330-1d7a2b80-3119-11e9-9ef4-c0542b7619bb.png)