Namerd stops watching dtabs if it receives a 404 during watch restart (#2223)
This PR is almost the same as #2192 but adds a fix to the exact piece of code that introduces this bug.

The implemented restartWatches currently has an issue where HTTP 404 responses from a k8s apiserver could cause the method to fail silently. This has the effect of stopping watch streams whenever HTTP 404 response is encountered.

This PR fixes the issue by pattern matching the response to a get request in the restartWatches method since it returns an Option[O] where O is the type of a k8s resource object. In the case where we receive a None the method needs to return a default empty value in order for the underlying Watchable object to restart the watch successfully.

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>