add support for HTTP/2 ping frame and add failure detector for mesh interpreter (#2221)
The current HTTP/2 implementation does not support ping frames between an H2 server and client. This makes it difficult to implement a failure detector in Linkerd or Namerd to ensure that connections are alive and in a good state.

To fix this issue, This PR, adds a "listener" to the existing implementation that notifies the transport when a PING frame is received on both the client and the server. This PR also implements a failure detector for HTTP/2 connections between Linkerd and Namerd which ensures that no stale connections between the two instance stay up longer than a given period of time.

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>