adds support for dynamically modifying log levels in Namerd (#2258)
Namerd's dashboard currently does not have the ability to dynamically change log levels for modules running within the Namerd process. This makes it difficult to debug issues that may arise in Namerd since you are left to rely on the current log levels set at startup time.

This PR adds a new "Logging" page which allows you to change log levels dynamically. This PR also includes a nav bar for Namerd to assist in switching between the dtab playground and the new logging page.

Fixes #2254

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>