Remove delegate tree from rich exceptions (#2261)
RichConnectionFailedException and RichNoBrokersException make a call to the `delegete` when they are populated to provide the delegate tree in the text of the exception.  In cases where there is a high rate of these exceptions being created (for example, when there are many requeues) this results in a high rate of calls to `delegate`.  If Namerd is used for delegation, this results in a high rate of requests to Namerd which can have Denial of Service effects.

We remove the delegation tree information from the rich exceptions.  This ensures that Namerd will not be impacted when there is a high rate of these exceptions.  The rich exceptions will still contain the service name and client name, so the delegation tree can still be retrieved manually from the dtab UI in the admin server if it is desired.

Signed-off-by: Alex Leong <alex@buoyant.io>